requirements.txt
ollama
xmpppy
